requirements.txt
argon2-cffi==21.1.0
blinker==1.4
cffi==1.14.6
click==8.0.1
colorama==0.4.4
Flask==1.1.2
Flask-Mail==0.9.1
Flask-WTF==0.15.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
neo4j==4.3.6
pycparser==2.20
pytz==2021.3
typing-extensions==3.10.0.2
Werkzeug==2.0.2
WTForms==2.3.3
zipp==3.6.0
